requirements.txt
cvxopt==1.3.0
gym==0.18.0
keras==2.8.0
matplotlib==3.3.2
nn_builder==0.0.09
numpy==1.21.5
pandas==1.1.2
pyod==0.9.9
PyYAML==6.0
scikit_learn==0.24.0
scipy==1.5.2
tensorflow==1.15.0
toml==0.10.2
torch==1.8.0
torchvision==0.7.0
tqdm==4.51.0
xgboost==1.6.1
keras==2.3.1